+++ doc/sphinx/changelog.rst
@@ -33,6 +33,7 @@
 
   * Improve the support and the API for ``LF_ENUM``: |lief-pdb-types-Enum|
   * Improve the support and the API for ``LF_PROCEDURE``: |lief-pdb-types-Function|
+  * Improve the support and the API for ``LF_ARRAY``: |lief-pdb-types-Array|
 
 :COFF: